pyproject.toml
[tool.poetry]
name = "referal"
version = "0.1.0"
description = ""
authors = ["deriabin_85 <deriabin_85@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
